fix multiple calls to emptyCache if the OS takes a while to delete the directory

+++ flutter_cache_manager/lib/src/storage/file_system/file_system_io.dart
@@ -7,13 +7,21 @@ import 'package:file/local.dart';
 import 'package:flutter_cache_manager/src/storage/file_system/file_system.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
+import 'package:uuid/uuid.dart';
 
 class IOFileSystem implements FileSystem {
   final Future<Directory> _fileDir;
   final bool _useIsolates;
 
-  IOFileSystem(Future<Directory> dir, {bool useIsolates = true})
-      : _useIsolates = useIsolates,
+  // Useful for testing, to mock slow deletes of big cache directories;
+  final Duration? _deleteDelay;
+
+  IOFileSystem(
+    Future<Directory> dir, {
+    bool useIsolates = true,
+    Duration? deleteDelay,
+  })  : _useIsolates = useIsolates,
+        _deleteDelay = deleteDelay,
         _fileDir = dir.then((value) => _createDir(value));
 
   factory IOFileSystem.fromCacheKey(String cacheKey) =>
@@ -47,19 +55,35 @@ class IOFileSystem implements FileSystem {
     final directory = await _fileDir;
 
     if (await directory.exists()) {
-      Future<void> handleDelete() async {
-        try {
-          final dirToDelete =
-              await directory.rename('${directory.path}.remove');
-          await dirToDelete.delete(recursive: true);
-        } on PathNotFoundException catch (_) {
-          // Avoid race conditions where the file might already be deleted by the OS
+      // It can take a while to delete the directory, so we rename it first
+      // and let it delete in the background.
+
+      Directory? dirToDelete;
+      try {
+        final dirToDeletePath = '${directory.path}.${const Uuid().v1()}.remove';
+        dirToDelete = await directory.rename(dirToDeletePath);
+      } on FileSystemException catch (e) {
+        // Avoid race conditions where the file might already be deleted by the OS
+        if (!_isPathNotFound(e)) {
+          rethrow;
         }
       }
 
-      // It can take a while to delete the directory, so we rename it first
-      // and let it delete in the background.
-      unawaited(_useIsolates ? Isolate.run(handleDelete) : handleDelete());
+      if (dirToDelete != null) {
+        unawaited(_handleInBg(() async {
+          try {
+            if (_deleteDelay != null) {
+              await Future.delayed(_deleteDelay!);
+            }
+            await dirToDelete?.delete(recursive: true);
+          } on FileSystemException catch (e) {
+            // Avoid race conditions where the file might already be deleted by the OS
+            if (!_isPathNotFound(e)) {
+              rethrow;
+            }
+          }
+        }));
+      }
     }
   }
 
@@ -67,20 +91,44 @@ class IOFileSystem implements FileSystem {
   Future<void> deleteDanglingCache() async {
     final directory = await _fileDir;
 
-    const fs = LocalFileSystem();
-    final dirToDelete = fs.directory('${directory.path}.remove');
+    final dirsToDelete = await directory.parent
+        .list()
+        .where((d) => d.path.endsWith('.remove'))
+        .toList();
 
-    if (await dirToDelete.exists()) {
-      Future<void> handleDelete() async {
-        try {
-          // print("Deleting cache dir: $dirToDelete");
-          await dirToDelete.delete(recursive: true);
-        } on PathNotFoundException catch (_) {
-          // Avoid race conditions where the file might already be deleted by the OS
+    if (dirsToDelete.isNotEmpty) {
+      await _handleInBg(() async {
+        final futures = <Future<void>>[];
+        for (final dirToDelete in dirsToDelete) {
+          futures.add(() async {
+            if (await dirToDelete.exists()) {
+              try {
+                // print("Deleting dangling cache dir: $dirToDelete");
+                if (_deleteDelay != null) {
+                  await Future.delayed(_deleteDelay!);
+                }
+                await dirToDelete.delete(recursive: true);
+              } on FileSystemException catch (e) {
+                // Avoid race conditions where the file might already be deleted by the OS
+                if (!_isPathNotFound(e)) {
+                  rethrow;
+                }
+              }
+            }
+          }());
         }
-      }
 
-      unawaited(_useIsolates ? Isolate.run(handleDelete) : handleDelete());
+        await Future.wait(futures);
+      });
     }
   }
+
+  Future<void> _handleInBg(Future<void> Function() fn) async {
+    await (_useIsolates ? Isolate.run(fn) : fn());
+  }
+}
+
+bool _isPathNotFound(FileSystemException e) {
+  return e is PathNotFoundException ||
+      e.osError?.errorCode == ErrorCodes.ENOENT;
 }
